+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Added
 - Added preflight sendout screen
 - Added more user group permissions
+- Added verified timestamps for GDPR compliance
+- Added purging of pending contacts with "purgePendingUsersDuration" config setting
 ### Changed
 - Changed "Settings" to "Plugin Settings"
 - Changed "userId" to "senderId" in Sendouts table